[CHORE]#263 : NotificationData model 패키지로 분리, 알림 타입에 따라 화면 전환 로직 NotificationNavigator로 변경
