+++ app/src/main/kotlin/com/bff/wespot/main/state/MainSideEffect.kt
@@ -1,7 +1,7 @@
 package com.bff.wespot.main.state
 
 import com.bff.wespot.main.model.VersionUpdateType
-import com.bff.wespot.notification.PushNotificationData
+import com.bff.wespot.model.notification.PushNotificationData
 
 sealed interface MainSideEffect {
     data class ShowVersionUpdateDialog(val versionUpdateType: VersionUpdateType) : MainSideEffect
