+++ core/navigation/build.gradle.kts
@@ -9,5 +9,6 @@ android {
 }
 
 dependencies {
+    implementation(project(":core:model"))
     implementation(libs.androidx.compose.destination)
 }
