+++ core/navigation/src/main/kotlin/com/bff/wespot/navigation/NotificationNavigator.kt
@@ -1,6 +1,10 @@
 package com.bff.wespot.navigation
 
+import android.content.Context
+import com.bff.wespot.model.notification.PushNotificationData
+
 interface NotificationNavigator {
+    fun navigate(context: Context, data: PushNotificationData)
     fun navigateUp()
     fun navigateToReceiverSelectionScreen()
     fun navigateToVotingScreen()
